@@ -66,8 +66,8 @@
     "glob": "^7.1.2",
     "jest": "^22.4.2",
     "path": "^0.12.7",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2",
     "tslint": "^5.9.1",
     "typescript": "^2.8.1",
     "webpack": "^4.18.0",
